+++ src/runtime/middleware/oidcAuth.ts
@@ -34,8 +34,8 @@ export default defineNuxtRouteMiddleware(async (to) => {
   if (loggedIn.value === true || to.path.startsWith('/auth/')) {
     return
   }
-  if (useRuntimeConfig().oidc?.middleware?.redirect === false) {
+  if ((useRuntimeConfig().oidc?.middleware as { redirect?: boolean } | undefined)?.redirect === false) {
     return
   }
-  await login()
+  await login(undefined, { callbackRedirectUrl: to.fullPath })
 })